requirements.txt
codename==1.1
django
django-filterpip==24.3
djangorestframework==3.15.2
djangorestframework-gis=1.1
drf-spectacular==0.28.0
drf-spectacular-sidecar=2025.5.1
python-dotenv==1.1.0